Add files via upload
A very simple encryption method that has been in use for thousands of years is 
called a Caesar cipher. Recall that the character set for text is ordered as a 
sequence of distinct values. This encryption strategy replaces each character in 
the plain text with the character that occurs a given distance away in the 
sequence. For positive distances, the method wraps around to the beginning of 
the sequence to locate the replacement characters for those characters near its 
end. For example, if the distance value of a Caesar cipher equals two characters, 
the string “california” would be encrypted as “ecnkhqtpkc”. To decrypt this cipher 
text back to plain text, you apply a method that uses the same distance value but 
looks to the left of each character for its replacement. This decryption method 
wraps around to the end of the sequence to find a replacement character for one 
near its beginning.